+++ assets/javascripts/kitten/components/buttons/button/helpers/modifier-styles.js
@@ -79,30 +79,28 @@ export const modifierStyles = modifier => {
     transition: background-color 0.2s, color 0.2s, border-color 0.2s;
 
     ${modifier !== 'checked' &&
-      css`
-        :hover,
-        :focus {
-          border-color: ${hoverBorderColor};
-          background-color: ${hoverBgColor};
-          color: ${hoverColor};
+    css`
+      :hover,
+      :focus {
+        border-color: ${hoverBorderColor};
+        background-color: ${hoverBgColor};
+        color: ${hoverColor};
 
-          svg {
-            fill: ${hoverColor};
-          }
+        svg {
+          fill: ${hoverColor};
         }
+      }
 
-        :active {
-          border-color: ${activeBorderColor};
-          background-color: ${activeBgColor};
-          color: ${activeColor};
+      :active {
+        border-color: ${activeBorderColor};
+        background-color: ${activeBgColor};
+        color: ${activeColor};
 
-          svg {
-            fill: ${activeColor};
-          }
-        }`
-    }
-
-    :disabled {
+        svg {
+          fill: ${activeColor};
+        }
+      }
+    `} :disabled {
       border-color: ${COLORS.line2};
       background-color: ${COLORS.line2};
       color: ${COLORS.background1};