+++ src/components/icons/search-icon/index.js
@@ -15,16 +15,25 @@ var _react = _interopRequireDefault(require("react"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
+var _deprecated = _interopRequireDefault(require("prop-types-extra/lib/deprecated"));
+
 var SearchIcon = function SearchIcon(_ref) {
   var iconTitle = _ref.iconTitle,
+      title = _ref.title,
       color = _ref.color,
       circleProps = _ref.circleProps,
       pathProps = _ref.pathProps,
-      other = (0, _objectWithoutProperties2.default)(_ref, ["iconTitle", "color", "circleProps", "pathProps"]);
+      width = _ref.width,
+      height = _ref.height,
+      props = (0, _objectWithoutProperties2.default)(_ref, ["iconTitle", "title", "color", "circleProps", "pathProps", "width", "height"]);
   return /*#__PURE__*/_react.default.createElement("svg", (0, _extends2.default)({
+    role: "img",
+    "aria-label": iconTitle || title,
+    xmlns: "http://www.w3.org/2000/svg",
     viewBox: "-0.7 -0.7 18 18",
-    xmlns: "http://www.w3.org/2000/svg"
-  }, other), iconTitle && /*#__PURE__*/_react.default.createElement("title", null, iconTitle), /*#__PURE__*/_react.default.createElement("circle", (0, _extends2.default)({
+    width: width,
+    height: height
+  }, props), (iconTitle || title) && /*#__PURE__*/_react.default.createElement("title", null, iconTitle || title), /*#__PURE__*/_react.default.createElement("circle", (0, _extends2.default)({
     cx: "7",
     cy: "7",
     r: "7",
@@ -39,14 +48,19 @@ var SearchIcon = function SearchIcon(_ref) {
 
 exports.SearchIcon = SearchIcon;
 SearchIcon.propTypes = {
-  iconTitle: _propTypes.default.string,
+  iconTitle: (0, _deprecated.default)(_propTypes.default.string, 'Prefere use `title` prop instead'),
+  title: _propTypes.default.string,
   color: _propTypes.default.string,
   circleProps: _propTypes.default.object,
-  pathProps: _propTypes.default.object
+  pathProps: _propTypes.default.object,
+  width: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string]),
+  height: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string])
 };
 SearchIcon.defaultProps = {
-  iconTitle: 'Search',
+  title: null,
   color: '#333',
   circleProps: {},
-  pathProps: {}
+  pathProps: {},
+  width: 14,
+  height: 14
 };